Revert "Command line --timeout option for server stop"
