+++ dev/com.ibm.ws.kernel.boot.core/src/com/ibm/ws/kernel/boot/internal/commands/ProcessControlHelper.java
@@ -110,18 +110,16 @@ public class ProcessControlHelper {
             stopRc = ReturnCode.REDUNDANT_ACTION_STATUS;
         }
 
-        int timeout = launchArgs.getStopTimeout();
-
         // If the lock file existed before we attempted to stop the server,
         // wait until we can obtain the server lock file (because the server process has stopped)
         if (stopRc == ReturnCode.OK && lockExists) {
-            stopRc = serverLock.waitForStop(timeout);
+            stopRc = serverLock.waitForStop();
         }
 
         if (stopRc == ReturnCode.OK) {
             String pid = getPID();
             if (pid != null) {
-                stopRc = waitForProcessStop(pid, timeout);
+                stopRc = waitForProcessStop(pid);
             }
         }
 
@@ -138,30 +136,22 @@ public class ProcessControlHelper {
         return stopRc;
     }
 
-    private ReturnCode waitForProcessStop(String pid, int timeout) {
+    private ReturnCode waitForProcessStop(String pid) {
         ProcessStatus ps = new PSProcessStatusImpl(pid);
 
-        final long timeoutMillis = timeout * 1000;
-        final long startTime = System.currentTimeMillis();
-        final long expireTime = startTime + timeoutMillis;
-
-        do {
+        for (int i = 0; i < BootstrapConstants.MAX_POLL_ATTEMPTS; i++) {
             try {
                 State processRunning = ps.isPossiblyRunning();
                 if ((processRunning == State.NO) || (processRunning == State.UNDETERMINED)) {
                     return ReturnCode.OK;
                 }
 
-                long timeNow = System.currentTimeMillis();
-                if ((timeNow) > expireTime) {
-                    break;
-                }
                 Thread.sleep(BootstrapConstants.POLL_INTERVAL_MS);
             } catch (Exception e) {
                 Debug.printStackTrace(e);
                 break;
             }
-        } while (true);
+        }
 
         return ReturnCode.ERROR_SERVER_STOP;
     }
