+++ dev/com.ibm.ws.kernel.boot.core/src/com/ibm/ws/kernel/boot/internal/commands/ServerHelpActions.java
@@ -28,7 +28,7 @@ public class ServerHelpActions implements HelpActions {
         startCmd(Category.lifecycle, "clean"),
         startWinServiceCmd(Category.win),
         statusCmd(Category.lifecycle),
-        stopCmd(Category.lifecycle, "force", "timeout"),
+        stopCmd(Category.lifecycle, "force"),
         stopWinServiceCmd(Category.win),
         unregisterWinServiceCmd(Category.win),
         versionCmd(Category.help);
