+++ dev/com.ibm.ws.kernel.boot_fat/fat/src/com/ibm/ws/kernel/boot/FATSuite.java
@@ -49,7 +49,6 @@ import com.ibm.wsspi.kernel.embeddable.EmbeddedServerTest;
                 KernelChangeTest.class,
                 ServerEnvTest.class,
                 ServerStartTest.class,
-                ServerStopTest.class,
                 ServerStartAsServiceTest.class,
                 ShutdownTest.class,
                 ServerCommandPortTest.class,
