+++ dev/com.ibm.ws.kernel.service/src/com/ibm/wsspi/kernel/service/utils/MetatypeUtils.java
@@ -12,13 +12,17 @@ package com.ibm.wsspi.kernel.service.utils;
 
 import java.util.Arrays;
 import java.util.Collection;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.Map;
 import java.util.concurrent.TimeUnit;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 import com.ibm.websphere.ras.Tr;
 import com.ibm.websphere.ras.TraceComponent;
 import com.ibm.websphere.ras.annotation.Trivial;
 import com.ibm.ws.ffdc.annotation.FFDCIgnore;
-import com.ibm.ws.kernel.boot.internal.KernelUtils;
 
 /**
  * Utility class to simplify retrieving values from injected config.
@@ -36,6 +40,53 @@ import com.ibm.ws.kernel.boot.internal.KernelUtils;
 public class MetatypeUtils {
     static final TraceComponent tc = Tr.register(MetatypeUtils.class);
 
+    private static final Map<String, TimeUnit> UNIT_DESCRIPTORS;
+    static {
+        HashMap<String, TimeUnit> units = new HashMap<String, TimeUnit>();
+
+        // We used to allow the duration abbreviations to be translated, but
+        // that causes config to be non-portable, so we now recommend that
+        // customers use English abbreviations only.  We hardcode the previously
+        // translated abbreviations for compatibility.  Additional translations
+        // should not be added.
+
+        units.put("d", TimeUnit.DAYS); // en, es, ja, ko, pl, pt_BR, zh
+        units.put("dn\u016f", TimeUnit.DAYS); // cs
+        units.put("g", TimeUnit.DAYS); // it
+        units.put("j", TimeUnit.DAYS); // fr
+        units.put("n", TimeUnit.DAYS); // hu
+        units.put("t", TimeUnit.DAYS); // de
+        units.put("z", TimeUnit.DAYS); // ro
+        units.put("\u0434", TimeUnit.DAYS); // ru
+        units.put("\u5929", TimeUnit.DAYS); // zh_TW
+
+        //units.put("g", TimeUnit.HOURS); // pl - conflicts with DAYS for "it"
+        units.put("h", TimeUnit.HOURS); // en, de, es, fr, it, ja, ko, pt_BR, ro, zh
+        units.put("hod", TimeUnit.HOURS); // cs
+        units.put("\u00f3", TimeUnit.HOURS); // hu
+        units.put("\u0447", TimeUnit.HOURS); // ru
+        units.put("\u5c0f\u6642", TimeUnit.HOURS); // zh_TW
+
+        units.put("m", TimeUnit.MINUTES); // en, de, es, fr, hu, it, ja, ko, pl, pt_BR, ro, zh
+        units.put("min", TimeUnit.MINUTES); // cs
+        units.put("\u043c", TimeUnit.MINUTES); // ru
+        units.put("\u5206", TimeUnit.MINUTES); // zh_TW
+
+        units.put("e", TimeUnit.SECONDS); // pt_BR
+        units.put("mp", TimeUnit.SECONDS); // hu
+        units.put("s", TimeUnit.SECONDS); // en, cs, de, es, fr, it, ja, ko, pl, ro, zh
+        units.put("\u0441", TimeUnit.SECONDS); // ru
+        units.put("\u79d2", TimeUnit.SECONDS); // zh_TW
+
+        units.put("ms", TimeUnit.MILLISECONDS); // en, cs, de, es, fr, hu, it, ja, ko, pl, pt_BR, ro, zh
+        units.put("\u043c\u0441", TimeUnit.MILLISECONDS); // ru
+        units.put("\u6beb\u79d2", TimeUnit.MILLISECONDS); // zh_TW
+
+        UNIT_DESCRIPTORS = Collections.unmodifiableMap(units);
+    }
+
+    private final static Pattern INTERVAL_STRING = Pattern.compile("(\\d+)(\\D+)");
+
     /**
      * Parse a boolean from the provided config value: checks for whether or not
      * the object read from the Service/Component configuration is a String
@@ -45,7 +96,7 @@ public class MetatypeUtils {
      * A translated warning message will be issued using the provided propertyKey and object
      * as parameters. FFDC for the exception is suppressed: Callers should handle the thrown
      * IllegalArgumentException as appropriate.
-     *
+     * 
      * @param configAlias
      *            Name of config (pid or alias) associated with a registered service
      *            or DS component.
@@ -54,7 +105,7 @@ public class MetatypeUtils {
      *            Used in the warning message if the value is badly formed.
      * @param obj
      *            The object retrieved from the configuration property map/dictionary.
-     *
+     * 
      * @return boolean parsed from obj, the default value if obj is null.
      * @throws IllegalArgumentException If value is not a String/Boolean, or if the String
      *             boolean is not "true" or "false" (ignoring case)
@@ -91,7 +142,7 @@ public class MetatypeUtils {
      * A translated warning message will be issued using the provided propertyKey and object
      * as parameters. FFDC for the exception is suppressed: Callers should handle the thrown
      * IllegalArgumentException as appropriate.
-     *
+     * 
      * @param configAlias
      *            Name of config (pid or alias) associated with a registered service
      *            or DS component.
@@ -144,7 +195,7 @@ public class MetatypeUtils {
      * A translated warning message will be issued using the provided propertyKey and object
      * as parameters. FFDC for the exception is suppressed: Callers should handle the thrown
      * IllegalArgumentException as appropriate.
-     *
+     * 
      * @param configAlias
      *            Name of config (pid or alias) associated with a registered service
      *            or DS component.
@@ -155,7 +206,7 @@ public class MetatypeUtils {
      *            The object retrieved from the configuration property map/dictionary.
      * @param defaultValue
      *            The default value that should be applied if the object is null.
-     *
+     * 
      * @return Collection of strings parsed/retrieved from obj, or default value if obj is null
      * @throws IllegalArgumentException If value is not a String, String collection, or String array, or if an error
      *             occurs while converting/casting the object to the return parameter type.
@@ -196,7 +247,7 @@ public class MetatypeUtils {
      * A translated warning message will be issued using the provided propertyKey and object
      * as parameters. FFDC for the exception is suppressed: Callers should handle the thrown
      * IllegalArgumentException as appropriate.
-     *
+     * 
      * @param configAlias
      *            Name of config (pid or alias) associated with a registered service
      *            or DS component.
@@ -207,7 +258,7 @@ public class MetatypeUtils {
      *            The object retrieved from the configuration property map/dictionary.
      * @param defaultValue
      *            The default value that should be applied if the object is null.
-     *
+     * 
      * @return Long parsed/retrieved from obj, or default value if obj is null
      * @throws IllegalArgumentException If value is not a String/Short/Integer/Long, or if an error
      *             occurs while converting/casting the object to the return parameter type.
@@ -246,7 +297,7 @@ public class MetatypeUtils {
      * A translated warning message will be issued using the provided propertyKey and object
      * as parameters. FFDC for the exception is suppressed: Callers should handle the thrown
      * IllegalArgumentException as appropriate.
-     *
+     * 
      * @param configAlias
      *            Name of config (pid or alias) associated with a registered service
      *            or DS component.
@@ -257,7 +308,7 @@ public class MetatypeUtils {
      *            The object retrieved from the configuration property map/dictionary.
      * @param defaultValue
      *            The default value that should be applied if the object is null.
-     *
+     * 
      * @return Integer parsed/retrieved from obj, or default value if obj is null
      * @throws IllegalArgumentException If value is not a String/Short/Integer/Long, or if an error
      *             occurs while converting/casting the object to the return parameter type.
@@ -296,7 +347,7 @@ public class MetatypeUtils {
      * A translated warning message will be issued using the provided propertyKey and object
      * as parameters. FFDC for the exception is suppressed: Callers should handle the thrown
      * IllegalArgumentException as appropriate.
-     *
+     * 
      * @param configAlias
      *            Name of config (pid or alias) associated with a registered service
      *            or DS component.
@@ -363,7 +414,7 @@ public class MetatypeUtils {
      * A translated warning message will be issued using the provided propertyKey and object
      * as parameters. FFDC for the exception is suppressed: Callers should handle the thrown
      * IllegalArgumentException as appropriate.
-     *
+     * 
      * @param configAlias
      *            Name of config (pid or alias) associated with a registered service
      *            or DS component.
@@ -374,7 +425,7 @@ public class MetatypeUtils {
      *            The object retrieved from the configuration property map/dictionary.
      * @param defaultValue
      *            The default value that should be applied if the object is null.
-     *
+     * 
      * @return Long parsed/retrieved from obj, or default value if obj is null
      * @throws IllegalArgumentException If value is not a Long, or if an error
      *             occurs while converting/casting the object to the return parameter type.
@@ -392,7 +443,7 @@ public class MetatypeUtils {
      * A translated warning message will be issued using the provided propertyKey and object
      * as parameters. FFDC for the exception is suppressed: Callers should handle the thrown
      * IllegalArgumentException as appropriate.
-     *
+     * 
      * @param configAlias
      *            Name of config (pid or alias) associated with a registered service
      *            or DS component.
@@ -406,7 +457,7 @@ public class MetatypeUtils {
      * @param units
      *            The unit of time for the duration value. This is only used when
      *            converting from a String value.
-     *
+     * 
      * @return Long parsed/retrieved from obj, or default value if obj is null
      * @throws IllegalArgumentException If value is not a String/Short/Integer/Long, or if an error
      *             occurs while converting/casting the object to the return parameter type.
@@ -416,7 +467,7 @@ public class MetatypeUtils {
         if (obj != null) {
             try {
                 if (obj instanceof String)
-                    return KernelUtils.evaluateDuration((String) obj, units);
+                    return evaluateDuration((String) obj, units);
                 else if (obj instanceof Long)
                     return (Long) obj;
             } catch (Exception e) {
@@ -434,22 +485,58 @@ public class MetatypeUtils {
 
     /**
      * Converts a string value representing a unit of time into a Long value.
-     *
+     * 
      * @param strVal
      *            A String representing a unit of time.
-     * @param endUnit
-     *            The unit of time into which the string value should be converted.
+     * @param unit
+     *            The unit of time that the string value should be converted into
      * @return Long The value of the string in the desired time unit
      */
+    @FFDCIgnore(NumberFormatException.class)
     public static Long evaluateDuration(String strVal, TimeUnit endUnit) {
-        return KernelUtils.evaluateDuration(strVal, endUnit);
+        // If the value is a number, simply return the numeric value as a long
+        try {
+            return Long.valueOf(strVal);
+        } catch (NumberFormatException ex) {
+            // ignore
+        }
+
+        // Otherwise, parse the duration with unit descriptors.
+        return evaluateDuration(strVal, endUnit, UNIT_DESCRIPTORS);
+    }
+
+    private static Long evaluateDuration(String strVal, TimeUnit endUnit, Map<String, TimeUnit> unitDescriptors) {
+        Matcher m = INTERVAL_STRING.matcher(strVal);
+        long retVal = 0;
+        boolean somethingParsed = false;
+        while (m.find()) {
+            somethingParsed = true;
+            // either of these could throw it's own Illegal argument exception
+            // if one of the component parts is bad.
+            Long numberVal = Long.valueOf(m.group(1));
+            String unitStr = m.group(2);
+            if (unitStr == null) {
+                throw new IllegalArgumentException("Could not parse configuration value as a duration: " + strVal);
+            }
+            TimeUnit sourceUnit = unitDescriptors.get(unitStr.trim().toLowerCase());
+            if (sourceUnit == null) {
+                throw new IllegalArgumentException("Could not parse configuration value as a duration: " + strVal);
+            }
+            retVal += endUnit.convert(numberVal, sourceUnit);
+        }
+
+        if (!somethingParsed) {
+            throw new IllegalArgumentException("Could not parse configuration value as a duration: " + strVal);
+        }
+
+        return retVal;
     }
 
     /**
      * Converts a String value into a token value by collapsing whitespace. All whitespace at the beginning and
      * end of the String will be removed, and contiguous sequences of whitespace will be replaced with a single
      * space character.
-     *
+     * 
      * @param strVal
      *            A String to be trimmed
      * @return String The collapsed String
